Moved UrlUploader component to its own directory.
