+++ packages/veritone-react-common/src/components/FilePicker/UrlUploader/UrlUploader.js
