+++ packages/veritone-react-common/src/components/FilePicker/UrlUploader/styles.scss
@@ -1,0 +1,83 @@
+@import 'src/styles/modules/variables';
+@import 'src/styles/modules/muiTypography';
+
+.urlUploader {
+    display: -webkit-flex;
+    display: flex;
+    -webkit-flex-direction: column;
+    flex-direction: column;
+    width:100%;
+    height: 100%;
+}
+
+.urlTextField {
+    @include mui-text('menu');
+    margin-top: 0;
+    margin-bottom: 5px;
+    width: 75%;
+    .fileUrlInputFocused {
+        color: #2597EC;
+    }
+    
+    .fileUrlInputInkbar:after {
+        background-color: #2597EC;
+    }
+}
+
+.fileUrlPickerInput {
+    margin-top: 13px;
+    input {
+        padding: 5px 0;
+    }
+}
+
+.urlUploaderInfoBox {
+    display: -webkit-flex;
+    display: flex;
+    -webkit-flex-direction: column;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+    background-color: $grey-1;
+    border: 4px dashed $grey-4;
+    padding-top: 15px;
+    padding-bottom: 15px;
+    width: 100%;
+    flex: auto;
+    text-align: center
+}
+
+.correctUrlText {
+    @include mui-text('body1');
+    margin-bottom: 15px;
+}
+
+.confirmLicenseText {
+    @include mui-text('caption');
+    color: $grey-5;
+}
+
+.imageContainer {
+    display: -webkit-box;
+    display: -moz-box;
+    display: -ms-flexbox;
+    display: -webkit-flex;
+    display: flex;
+
+    margin-top: 45px;
+    padding: 10px;
+    background-color: #777777;
+}
+
+.fileImage {
+    flex: 1;
+    display: flex;
+    flex-direction: column;
+    max-width: 100%;
+    img {
+        align-self: center;
+        width: auto;
+        height: 100%;
+        display: block;
+    }
+}
